Added Blog:Comment - verify comment password

+++ symfony/src/Controller/Front/BlogController.php
@@ -3,7 +3,8 @@
 namespace App\Controller\Front;
 
 use App\Form\ArticleType;
-use App\Form\CommentType;
+use App\Form\ArticleCommentType;
+use App\Form\ArticleCommentVerifyType;
 use App\Entity\Article;
 use App\Entity\ArticleComment;
 use App\Event\FlashEvent;
@@ -16,14 +17,15 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * @author yuu2dev
- * @todo 댓글 삭제
- * updated 2020.08.11
+ * @todo 댓글 검증 삭제 수정
+ * updated 2020.08.20
  */
 class BlogController extends AbstractController {
 
@@ -35,7 +37,7 @@ class BlogController extends AbstractController {
   /**
    * 블로그 게시물 일람
    * @Route("/blog", name="blog_article_index", methods={"GET"})
-   * @Template("front/Blog/index.twig")
+   * @Template("front/blog/index.twig")
    * @access public
    * @param Request $request
    * @param BlogService $blogService
@@ -54,20 +56,33 @@ class BlogController extends AbstractController {
 
   /**
    * 블로그 게시물 상세
-   * @Route("/blog/article/{id}", name="blog_article_show", methods={"GET"}, requirements={"id":"\d+"})
+   * @Route("/blog/article/{id}", name="blog_article_show", methods={"GET", "POST"}, requirements={"id":"\d+"})
    * @Entity("article", expr="repository.findArticleById(id)")
-   * @Template("front/Blog/show.twig")
+   * @Template("front/blog/show.twig")
    * @access public
    * @param Article $article
    * @param BlogService $blogService
    * @param EventDispatcherInterface $eventDispatcher
+   * @param Request $request
    * @return array
    */
-  public function article_show(?Article $article, BlogService $blogService, EventDispatcherInterface $eventDispatcher): array {
+  public function article_show(?Article $article, BlogService $blogService, EventDispatcherInterface $eventDispatcher, Request $request) {
     
     if (!$article) {
       $eventDispatcher->dispatch(FlashEvent::BLOG_ARTICLE_INVISIBLE);
       $eventDispatcher->dispatch(RedirectEvent::REDIRECT, new RedirectEvent);
+    }
+    
+    $form = $this->createForm(ArticleCommentType::class, new ArticleComment, ['method' => 'post']);
+    $form->handleRequest($request);
+    
+    if ($form->isSubmitted() && $form->isValid()) {
+      /** @var ArticleComment */
+      $comment = $form->getData();
+      $comment->setArticle($article);
+      
+      $blogService->writeComment($comment) ? $eventDispatcher->dispatch(FlashEvent::BLOG_ARTICLE_COMMENT_WRITE_SUCCESS) : $eventDispatcher->dispatch(FlashEvent::BLOG_ARTICLE_COMMENT_WRITE_FAILED);
+      return $this->redirect($request->getUri());
     }
 
     return [
@@ -76,13 +91,13 @@ class BlogController extends AbstractController {
       'Comment_cnt' => $blogService->countCommentsByEntity($article),
       'Categories' => $blogService->findCategories(),
       'RecentArticles' => $blogService->recentArticles(10),
-      'RecentTags' => $blogService->recentTags(30)
+      'RecentTags' => $blogService->recentTags(30),
+      'CommentForm' => $form->createView(),
     ];
   }
 
   /**
    * 블로그 게시물 작성 및 수정
-   * @todo CKEditor 이미지 업로드 기능
    * @Route("/blog/article/new", name="blog_article_new", methods={"GET", "POST"})
    * @Route("/blog/article/edit/{id}", name="blog_article_edit", methods={"GET", "PUT"}, requirements={"id":"\d+"})
    * @Template("front/blog/form.twig")
@@ -96,7 +111,7 @@ class BlogController extends AbstractController {
    */
   public function article_form(?Article $article, Request $request, BlogService $blogService, EventDispatcherInterface $eventDispatcher, TranslatorInterface $translator) {
     
-    $form = $article ? $this->createForm(ArticleType::class, $article, ['method' => 'PUT']) : $this->createForm(ArticleType::class, new Article);
+    $form = $article ? $this->createForm(ArticleType::class, $article, ['method' => 'put']) : $this->createForm(ArticleType::class, new Article);
     $form->handleRequest($request);
 
     if ($form->isSubmitted() && $form->isValid()) {
@@ -123,7 +138,7 @@ class BlogController extends AbstractController {
 
   /**
    * 블로그 게시물 삭제
-   * @Route("/blog/article/delete/{id}", name="blog_article_del", methods={"GET"}, requirements={"id":"\d+"})
+   * @Route("/blog/article/del/{id}", name="blog_article_del", methods={"GET"}, requirements={"id":"\d+"})
    * @access public
    * @param Article $article
    * @param BlogService $blogService
@@ -139,46 +154,8 @@ class BlogController extends AbstractController {
   }
 
   /**
-   * 블로그 댓글 작성
-   * @Route("/blog/comment/new", name="blog_comment_new", methods={"GET", "POST"})
-   * @Template("front/blog/comment/new.twig")
-   * @access public
-   * @param int $article_id
-   * @param BlogService $blogService
-   * @param EventDispatcherInterface $eventDispatcher
-   * @param Request $request
-   * @return array|object
-   */
-  public function comment_new(BlogService $blogService, EventDispatcherInterface $eventDispatcher, Request $request) {
-    
-    $form = $this->createForm(CommentType::class, new ArticleComment, ['action' => $this->generateUrl('blog_comment_new'), 'method' => 'POST']);
-    
-    $form->handleRequest($request);
-    
-    if ($form->isSubmitted() && $form->isValid()) {
-
-      $comment = $form->getData();
-      $article = $form->get('article')->getData();
-      $article = $blogService->findArticle($article);
-      
-      /**
-       * @todo 세션 플래시 메시지
-       */
-      $blogService->writeComment($comment->setArticle($article)) ? $eventDispatcher->dispatch(FlashEvent::BLOG_ARTICLE_COMMENT_WRITE_SUCCESS) : $eventDispatcher->dispatch(FlashEvent::BLOG_ARTICLE_COMMENT_WRITE_FAILED);
-
-      return $this->redirectToRoute('blog_article_show', ['id' => $article->getId()]);
-    }
-
-    return [
-      'form' => $form->createView()
-    ];
-  }
-
-  /**
-   * @todo
    * 블로그 댓글 수정
    * @Route("/blog/comment/edit/{id}", name="blog_comment_edit", methods={"GET", "PUT"}, requirements={"id":"\d+"})
-   * @Template("front/blog/comment.twig")
    */
   public function comment_edit(ArticleComment $comment, BlogService $blogService, EventDispatcherInterface $eventDispatcher, Request $request) {
     // $form = $this->createForm(CommentType::class, $comment, ['action' => $this->generateUrl('blog_comment_new'), 'method' => 'PUT']);
@@ -186,19 +163,57 @@ class BlogController extends AbstractController {
 
   /**
    * 블로그 댓글 삭제
-   * @Route("/blog/article/comment/delete/{id}", name="blog_comment_del", methods={"DELETE"}, requirements={"id":"\d+"})
+   * @Route("/blog/article/comment/del/{id}", name="blog_comment_del", methods={"GET", "DELETE"}, requirements={"id":"\d+"})
    * @access public
    * @param ArticleComment $comment
    * @param BlogService $blogService
    * @param EventDispatcherInterface $eventDispatcher
    * @param Request $request
    */
-  public function comment_del(?ArticleComment $comment, BlogService $blogService, EventDispatcherInterface $eventDispatcher, Request $request) {
+  public function comment_del(ArticleComment $comment, BlogService $blogService, EventDispatcherInterface $eventDispatcher, Request $request) {
+      
+    $response = new Response;
+
+    /**
+     * @todo 익명 세션 검증
+     */
+    if ($this->getUser() == $comment->getUser()) {
     
-    $referer = $request->headers->get('referer');
+      $blogService->removeComment($comment) ? $eventDispatcher->dispatch(FlashEvent::BLOG_ARTICLE_COMMENT_DEL_SUCCESS) : $eventDispatcher->dispatch(FlashEvent::BLOG_ARTICLE_COMMENT_DEL_FAILED);
+    }
+
+    return $this->redirectToRoute('blog_article_show', ['id' => $comment->getArticle()->getId()]);
+  }
+
+  /**
+   * @todo
+   * 블로그 댓글 검증
+   * @Route("/blog/article/comment/verify/{id}", name="blog_comment_verify", methods={"GET","POST"}, requirements={"id":"\d+"})
+   * @param ArticleComment $comment
+   * @param BlogService $blogService
+   * @param EventDispatcherInterface $eventDispatcher
+   * @param Request $request
+   */
+  public function comment_verify(ArticleComment $comment, BlogService $blogService, EventDispatcherInterface $eventDispatcher, Request $request) {
+
+    $response = new Response;
+
+    $form = $this->createForm(ArticleCommentVerifyType::class, new ArticleComment, ['method' => 'post']);
+    $form->handleRequest($request);
     
-    $blogService->removeComment($comment) ? $eventDispatcher->dispatch(FlashEvent::BLOG_ARTICLE_COMMENT_DEL_SUCCESS) : $eventDispatcher->dispatch(FlashEvent::BLOG_ARTICLE_COMMENT_DEL_FAILED);
+    if ($form->isSubmitted() && $form->isValid()) {
+      $password = $request->request->get('password');
+      
+      /**
+       * @todo 세션등록
+       */
+      // password_verify($password, $comment->getPassword()) ? $respose->setStatusCode(Response::HTTP_OK) : $response->setStatusCode(Response::HTTP_FORBIDDEN);
+    }
     
-    $this->redirect($referer);
+    $response->setContent(
+      $this->render('front/blog/comment/form_verify.twig', ['form' => $form->createView()])->getContent()
+    );
+
+    return $response;
   }
 }
